# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type GoalStatus = (typeof GoalStatus)[keyof typeof GoalStatus];

export const GoalStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Goal {
  id: number;
  userId: string;
  title: string;
  description: string;
  category: string;
  city: string;
  status: GoalStatus;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorImage?: string | null;
}

export interface GoalResponse {
  goal: Goal;
}

export interface GoalListResponse {
  goals: Goal[];
}

export interface GoalUpdate {
  id: number;
  goalId: number;
  userId: string;
  content: string;
  createdAt: string;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorImage?: string | null;
}

export interface GoalDetailResponse {
  goal: Goal;
  updates: GoalUpdate[];
}

export interface CreateGoalRequest {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  city?: string;
}

export type UpdateGoalRequestStatus =
  (typeof UpdateGoalRequestStatus)[keyof typeof UpdateGoalRequestStatus];

export const UpdateGoalRequestStatus = {
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface UpdateGoalRequest {
  status?: UpdateGoalRequestStatus;
  title?: string;
  description?: string;
}

export interface GoalUpdateResponse {
  update: GoalUpdate;
}

export interface GoalUpdateListResponse {
  updates: GoalUpdate[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListGoalsParams = {
  category?: string;
  city?: string;
};