# lib/api-zod/src/generated/types/createGoalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGoalRequest {
  /** @minLength 1 */
  title: string;
  description: string;
  category: string;
  city?: string;
}