# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalStatus } from "./goalStatus";

export interface Goal {
  id: number;
  userId: string;
  title: string;
  description: string;
  category: string;
  city: string;
  status: GoalStatus;
  createdAt: Date;
  updatedAt: Date;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorImage?: string | null;
}