# lib/api-zod/src/generated/types/goalUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalUpdate {
  id: number;
  goalId: number;
  userId: string;
  content: string;
  createdAt: Date;
  /** @nullable */
  authorName?: string | null;
  /** @nullable */
  authorImage?: string | null;
}