# lib/api-zod/src/generated/types/goalUpdateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GoalUpdate } from "./goalUpdate";

export interface GoalUpdateResponse {
  update: GoalUpdate;
}