# lib/api-zod/src/generated/types/mobileTokenExchangeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}