# lib/api-zod/src/generated/types/updateGoalRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Skyline API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateGoalRequestStatus } from "./updateGoalRequestStatus";

export interface UpdateGoalRequest {
  status?: UpdateGoalRequestStatus;
  title?: string;
  description?: string;
}